# OogstBeoordelingsAPI/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Demo API",
    "version": "v1"
  },
  "paths": {
    "/api/Assessor/LinkToHarvest": {
      "post": {
        "tags": [
          "Assessor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Assessor/SubmitReview": {
      "post": {
        "tags": [
          "Assessor"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Id",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "WeightKG",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Description",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "OptionalAdditionalInfo",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "MarkPresentation",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "MarkMoisture",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "MarkTaste",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Assessor/GetPendingHarvest": {
      "get": {
        "tags": [
          "Assessor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderedHarvestListDto"
              }
            }
          }
        }
      }
    },
    "/api/Assessor/GetToBeReviewedHarvest": {
      "get": {
        "tags": [
          "Assessor"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Assessor/GetClosedHarvest": {
      "get": {
        "tags": [
          "Assessor"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Grower/CreateHarvest": {
      "post": {
        "tags": [
          "Grower"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "File",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "WeightKG",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "SubType",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "HarvestDate",
            "required": true,
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "Type",
            "required": true,
            "type": "integer",
            "format": "int32",
            "enum": [
              1,
              2
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Grower/GetHarvest/{id}": {
      "get": {
        "tags": [
          "Grower"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetHarvestGrowerDto"
            }
          }
        }
      }
    },
    "/api/Grower/GetAllClosedHarvest": {
      "get": {
        "tags": [
          "Grower"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetHarvestGrowerDto"
              }
            }
          }
        }
      }
    },
    "/api/Grower/GetAllNotClosedHarvest": {
      "get": {
        "tags": [
          "Grower"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetHarvestGrowerDto"
              }
            }
          }
        }
      }
    },
    "/api/User/Login": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "UserName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/CreateAccount": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Username",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Password",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "EmailAddress",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "UserRole",
            "required": true,
            "type": "integer",
            "format": "int32",
            "enum": [
              1,
              2,
              3
            ]
          },
          {
            "in": "formData",
            "name": "FirstName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "LastName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Zipcode",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "City",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Adres",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/DeleteUser/{id}/{username}": {
      "post": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "GetHarvestAssessorDto": {
      "required": [
        "address",
        "growerId",
        "growerName",
        "harvestDate",
        "harvestId",
        "status",
        "subType",
        "type",
        "weightKG"
      ],
      "type": "object",
      "properties": {
        "file": {
          "format": "byte",
          "type": "string"
        },
        "harvestId": {
          "minLength": 1,
          "type": "string"
        },
        "weightKG": {
          "format": "int32",
          "type": "integer"
        },
        "subType": {
          "minLength": 1,
          "type": "string"
        },
        "harvestDate": {
          "format": "date-time",
          "type": "string"
        },
        "type": {
          "format": "int32",
          "enum": [
            1,
            2
          ],
          "type": "integer"
        },
        "status": {
          "format": "int32",
          "enum": [
            1,
            2,
            3,
            4
          ],
          "type": "integer"
        },
        "address": {
          "minLength": 1,
          "type": "string"
        },
        "growerName": {
          "minLength": 1,
          "type": "string"
        },
        "growerId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "GetHarvestGrowerDto": {
      "required": [
        "harvestDate",
        "harvestId",
        "subType",
        "type",
        "weightKG"
      ],
      "type": "object",
      "properties": {
        "file": {
          "format": "byte",
          "type": "string"
        },
        "harvestId": {
          "minLength": 1,
          "type": "string"
        },
        "weightKG": {
          "format": "int32",
          "type": "integer"
        },
        "subType": {
          "minLength": 1,
          "type": "string"
        },
        "harvestDate": {
          "format": "date-time",
          "type": "string"
        },
        "type": {
          "format": "int32",
          "enum": [
            1,
            2
          ],
          "type": "integer"
        },
        "status": {
          "format": "int32",
          "enum": [
            1,
            2,
            3,
            4
          ],
          "type": "integer"
        }
      }
    },
    "OrderedHarvestListDto": {
      "type": "object",
      "properties": {
        "harvestList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetHarvestAssessorDto"
          }
        },
        "nameHarvester": {
          "type": "string"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "type": "string"
        }
      }
    },
    "User": {
      "required": [
        "adres",
        "city",
        "emailAddress",
        "firstName",
        "lastName",
        "password",
        "username",
        "userRole",
        "zipcode"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        },
        "emailAddress": {
          "minLength": 1,
          "type": "string"
        },
        "userRole": {
          "format": "int32",
          "enum": [
            1,
            2,
            3
          ],
          "type": "integer"
        },
        "firstName": {
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "minLength": 1,
          "type": "string"
        },
        "zipcode": {
          "minLength": 1,
          "type": "string"
        },
        "city": {
          "minLength": 1,
          "type": "string"
        },
        "adres": {
          "minLength": 1,
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "basic",
      "description": "enter JWT bearer token"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}